Generate curvesplots per-cluster to PDF

This function generates a PDF file featuring dose-response curves for each cluster of deregulated genes. Each plot is labeled with the cluster ID and the number of transcripts in the cluster. By default, the curves are color-coded to indicate whether the trend is increasing, decreasing, U-shaped, or bell-shaped. Users can also apply any modifications available in the original DRomics::curvesplot() function.

Usage

curves_to_pdf(
  lonelyfishing_data,
  bmdboot_data,
  clustrfusion_data,
  id_col_for_curves = "id",
  tested_doses,
  annot_order,
  ...,
  xunit,
  xtitle,
  ytitle,
  colors = c(inc = "#1B9E77", dec = "#D95F02", U = "#7570B3", bell = "#E7298A"),
  path,
  output_filename,
  overwrite = TRUE
)

Arguments

lonelyfishing_data

The named list output of the lonelyfishing() function.

bmdboot_data

The DRomics bmdboot dataframe results after DRomics::bmdfilter()

clustrfusion_data

The named list output of the clustrfusion() function.

id_col_for_curves

The column giving the identification of each curve (default is "id")

tested_doses

A vector of the tested doses that can be found in the output of the DRomics::drcfit()function (unique(f$omicdata$dose))

annot_order

A vector specifying the prioritized order of annotation sources used in the clustrenrich() function. This order determines which term or pathway will be used as primary descriptor for each cluster. The relevance of each source can be based on the quantity or quality of information they provide, tailored to the specific case study. For example, for Danio rerio, you might prioritize sources like c("GO:BP", "KEGG", "WP") based on the abundance of information they offer.

...

Additional arguments passed to the DRomics::curvesplot() function

xunit

Unit for the x scale

xtitle

X-axis title

ytitle

Y-axis title

colors

A vector of colors for different trends (default is set at c("inc" = "#1B9E77", "dec" = "#D95F02", "U" = "#7570B3", "bell" = "#E7298A"))

path

Destination folder for the output data results.

output_filename

Output PDF filename.

overwrite

If TRUE, the function overwrites existing output files; otherwise, it reads the existing file. (default is set to FALSE).

Value

No return value; the .pdf file is downloaded and stored in the specified director. It holds a curvesplot per slide for each cluster, and those for the Friendly (if created), Lonely and All deregulated gene transcripts.