Generate a workflow summary CSV file

This function creates a concise summary table capturing key details derived from the workflow results from the lonelyfishing() function. It is formatted with each row being a combination of transcript, cluster and biological function annotation.

Usage

results_to_csv(
  lonelyfishing_data,
  bmdboot_data,
  path,
  output_filename,
  overwrite = TRUE
)

Arguments

lonelyfishing_data

The named list output of the lonelyfishing() function.

bmdboot_data

The DRomics bmdboot dataframe results after DRomics::bmdfilter()

path

Destination folder for the output data results.

output_filename

Output CSV filename.

overwrite

If TRUE, the function overwrites existing output files; otherwise, it reads the existing file. (default is set to FALSE).

Value

No return value; the .csv file is downloaded and stored in the specified director. It holds the results of the final step of the workflow with each row being a combination of transcript, gene and other identifiers, cluster and biological function annotation. This is to be used as a support for exploration and mechanism discovery.