Filter clusters based on their gene-set size

This function filters clusters based on size, selecting genes that belong to clusters meeting the user-defined size criteria.

Usage

clustrfiltr(getclustrs_data, size_filtr = 3)

Arguments

getclustrs_data

A dataframe of type t that typically corresponds to the output of the getclustrs() function. This input holds at least the columns named gene_id and clustr respectively holding Ensembl gene and cluster identifiers for the deregulated genes.

size_filtr

The minimum number of genes required for a cluster to be retained (by default: 3).

Value

A named list holding 2 components, where : - kept is a dataframe of type t similar to the getclustrs_data dataframe input with the rows kept after the filter (the genes are part of a cluster equal and over the size limit) -removed is a dataframe of type t similar to the getclustrs_data dataframe input with the rows removed after the filter (the genes are part of clusters under the size limit) - params is a list of the main parameters used